debug confing in advanced
