+++ documents/configuration/test.advancedExecutables.md
@@ -60,6 +60,7 @@ If it is an object it can contains the following properties:
 | `strictPattern`             | Test loading fails if one of the files matched by `pattern` is not a test executable. (Helps noticing unexpected crashes/problems under test loading.)                                                                                                                                                                                                                                                        |
 | `markAsSkipped`             | If true then all the tests related to the pattern are skipped. They can be run manually though.                                                                                                                                                                                                                                                                                                               |
 | `executableCloning`         | If enabled it creates a copy of the test executable before listing or running the tests. NOTE: discovery (`--help`) still uses the original file.                                                                                                                                                                                                                                                             |
+| `debug.configTemplate`      | Sets the necessary debug configurations and the debug button will work.                                                                                                                                                                                                                                                                                                                                       |
 | `executableSuffixToInclude` | Filter files based on suffix for faster discovery.                                                                                                                                                                                                                                                                                                                                                            |
 | `waitForBuildProcess`       | Prevents the extension of auto-reloading. With this linking failure might can be avoided. Can be true to use a default pattern that works for most cases, or a string to pass your own search pattern (regex) for processes.                                                                                                                                                                                  |
 | `catch2`                    | Object with framework specific settings. [Detail](https://github.com/matepek/vscode-catch2-test-adapter/blob/master/documents/configuration/test.advancedExecutables.md#framework-specific-settings)                                                                                                                                                                                                          |
